fix pagination

+++ internal/repository/roles_repository.go
@@ -268,7 +268,7 @@ type PermissionSource struct {
 	IsDirectly  bool             `db:"is_directly_assigned" json:"is_directly_assigned"`
 }
 
-func ListUserPermissionsSource(ctx context.Context, dbx bob.Executor, userId uuid.UUID, limit int, offset int) ([]PermissionSource, error) {
+func ListUserPermissionsSource(ctx context.Context, dbx bob.Executor, userId uuid.UUID, limit int64, offset int64) ([]PermissionSource, error) {
 	q := psql.RawQuery(QueryUserPermissionSource, userId, userId, limit, offset)
 
 	data, err := bob.All(ctx, dbx, q, scan.StructMapper[PermissionSource]())
@@ -369,7 +369,7 @@ WHERE cp.id IS NULL;
 ;`
 )
 
-func ListUserNotPermissionsSource(ctx context.Context, dbx bob.Executor, userId uuid.UUID, limit int, offset int) ([]PermissionSource, error) {
+func ListUserNotPermissionsSource(ctx context.Context, dbx bob.Executor, userId uuid.UUID, limit int64, offset int64) ([]PermissionSource, error) {
 	q := psql.RawQuery(getuserNotPermissions, userId, userId, limit, offset)
 
 	res, err := bob.All(ctx, dbx, q, scan.StructMapper[PermissionSource]())
